+++ src/frontend/web/src/components/Main/Aside/index.tsx
@@ -9,7 +9,7 @@ import ImageButton from '../../Common/ImageButton';
 import ArrowLeftIcon from '../../../icon/arrowLeft.svg';
 import { WorkspaceName } from '../../Common/WorkspaceName';
 import ModalMenuHook from '../../../hook/ModalMenu';
-import ModalMain from '../../Common/ModalMain';
+import ModalMain from '../ModalMain';
 
 const ModalBox = styled.div`
   display: flex;
